Game of War (Original)
This was my midterm project for my MSDS bridge course in python.
When this project is run, there are three modes to play the game:
The first is within the terminal.
The second is a simulation where the user inputs the number of times the game should run (This is not an optimized simulation by any means and the number should probably not be over 100 as you may die before it completes)
Finally the third mode opens the game on the tkinter canvas.